revert for now

+++ ext/standard/basic_functions.c
@@ -17,7 +17,7 @@
    +----------------------------------------------------------------------+
  */
 
-/* $Id: basic_functions.c,v 1.924 2008/08/20 22:19:57 pajoye Exp $ */
+/* $Id: basic_functions.c,v 1.925 2008/08/21 08:37:01 tony2001 Exp $ */
 
 #include "php.h"
 #include "php_streams.h"
@@ -4929,7 +4929,7 @@ PHP_FUNCTION(get_current_user)
  */
 static int add_config_entry_cb(zval *entry TSRMLS_DC, int num_args, va_list args, zend_hash_key *hash_key)
 {
-	zval *retval = (zval *)(long)va_arg(args, int);
+	zval *retval = (zval *)va_arg(args, int);
 	zval *tmp;
 
 	if (Z_TYPE_P(entry) == IS_STRING) {
